Updated index.js to contain all subgraphs for Workflow Facade, Internal Audit, and Notifications as well as demo graphs user, location, and feedback
